fix(gateway): validate Lazer cache freshness config (ENG-398)
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
